refactor[#501] : 이 장소 마킹 바텀 시트 모듈화 리팩토링
