+++ src/widgets/map/ui/index.ts
@@ -5,6 +5,6 @@ export * from "./mapMarkerWidget";
 export * from "./mapInitializer";
 export * from "./mapBottomSheet";
 export * from "./localMarkingList";
-export * from "./placeMarkingList";
+export * from "./placeMarkingListBottomSheet";
 export * from "./myMarkingList";
 export * from "./myActivityList";
